Remove special case node.js handling of functions with callback in the name.
This feature seemed very specific to ledgers use case, is implemented in a 'hacky' sort of way (based on method names instead of marked in the IDL) and doesn't seem suited to something that other projects might use.
So we remove it to simplify things, if the functionality is ever actually needed it should be added back in a more fleshed out way.