Introduce new '&' return type to IDL for "maybe async" functionality.
What:
When marked with this return type generators will generate the interface method as returning void
On platforms that require it (currently only node) the generator will generate the function as a less immediate and slower async call (e.g. via event loop) instead of a direct call
On platforms that don't require it the generator will still generate this as a direct and more efficient call

Why:
On node.js (electron) it is not legal to call back into the V8 engine (javascript) from C++ if you aren't currently in the main thread.
This restriction is fine if your code works as follows <electron -> blocking_djinni_backend_call -> work -> javascript_interface_calls> - which is exactly what people may want for simpler short workloads, for longer workloads or something complex like what we have at gulden (framework with a permanently running event loop) it is not.
It is problematic (crashes) if your code works as follows <electron -> djinni_call_that_starts_work_in_thread>   <worker_thread -> javascript_interface_calls> - which crashes

As "fixing" this adds overhead and means less immediate calls with higher latency it isn't appropriate to always "fix" this.
Therefore the new syntax allows binding maintainers to selectively choose which functions should be available for which type of workload - and only incur the overhead when they want it.